chebyshev mesh + refactoring
